requirements.txt
streamlit>=1.28.0
openai>=1.0.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
pydantic>=2.0.0
python-dotenv>=1.0.0
pypdf2>=3.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
